// src/exam/interfaces/toxicologicalSample.interface.ts
export interface IToxicologicalSample {
  codigo_amostra: string;
  Cocaína: number;
  Anfetamina: number;
  Metanfetamina: number;
  MDA: number;
  MDMA: number;
  THC: number;
  Morfina: number;
  Codeína: number;
  Heroína: number;
  Benzoilecgonina: number;
  Cocaetileno: number;
  Norcocaína: number;
}